+++ components/builder-api/src/lib.rs
@@ -30,6 +30,7 @@ extern crate diesel_migrations;
 #[macro_use]
 extern crate features;
 extern crate github_api_client;
+extern crate habitat_builder_db as db;
 extern crate habitat_builder_protocol as protocol;
 extern crate habitat_core as hab_core;
 extern crate habitat_http_client as http_client;
