+++ components/builder-api/src/server/mod.rs
@@ -13,10 +13,11 @@
 // limitations under the License.
 
 pub mod authorize;
+pub mod db;
 pub mod error;
 pub mod framework;
 pub mod helpers;
-pub mod models;
+pub mod model;
 pub mod resources;
 pub mod schema;
 pub mod services;
@@ -36,12 +37,9 @@ use hab_net::socket;
 use oauth_client::client::OAuth2Client;
 use segment_api_client::SegmentClient;
 
+use self::db::{init, DbPool};
 use self::error::Error;
 use self::framework::middleware::{Authentication, Cors, XRouteClient};
-use self::services::route_broker::RouteBroker;
-use self::services::s3::S3Handler;
-use self::services::upstream::{UpstreamClient, UpstreamMgr};
-
 use self::resources::authenticate::Authenticate;
 use self::resources::channels::Channels;
 use self::resources::ext::Ext;
@@ -52,6 +50,9 @@ use self::resources::pkgs::Packages;
 use self::resources::profile::Profile;
 use self::resources::projects::Projects;
 use self::resources::user::User;
+use self::services::route_broker::RouteBroker;
+use self::services::s3::S3Handler;
+use self::services::upstream::{UpstreamClient, UpstreamMgr};
 
 use config::{Config, GatewayCfg};
 
@@ -71,6 +72,7 @@ pub struct AppState {
     oauth: OAuth2Client,
     segment: SegmentClient,
     upstream: UpstreamClient,
+    db: DbPool,
 }
 
 impl AppState {
@@ -82,6 +84,7 @@ impl AppState {
             oauth: OAuth2Client::new(config.oauth.clone()),
             segment: SegmentClient::new(config.segment.clone()),
             upstream: UpstreamClient::default(),
+            db: init(config.datastore.clone()),
         }
     }
 }
