--- components/builder-api/src/server/schema.rs
@@ -1,145 +1,0 @@
-table! {
-    origin_channel_packages (channel_id, package_id) {
-        channel_id -> BigInt,
-        package_id -> BigInt,
-        created_at -> Nullable<Timestamptz>,
-        updated_at -> Nullable<Timestamptz>,
-    }
-}
-
-table! {
-    origin_channels (id) {
-        id -> BigInt,
-        origin_id -> BigInt,
-        owner_id -> BigInt,
-        name -> Text,
-        created_at -> Nullable<Timestamptz>,
-        updated_at -> Nullable<Timestamptz>,
-    }
-}
-
-table! {
-    origin_integrations (id) {
-        id -> BigInt,
-        origin -> Nullable<Text>,
-        integration -> Nullable<Text>,
-        name -> Nullable<Text>,
-        body -> Nullable<Text>,
-        created_at -> Nullable<Timestamptz>,
-        updated_at -> Nullable<Timestamptz>,
-    }
-}
-
-table! {
-    origin_invitations (id) {
-        id -> BigInt,
-        origin_id -> Nullable<BigInt>,
-        origin_name -> Nullable<Text>,
-        account_id -> Nullable<BigInt>,
-        account_name -> Nullable<Text>,
-        owner_id -> Nullable<BigInt>,
-        ignored -> Nullable<Bool>,
-        created_at -> Nullable<Timestamptz>,
-        updated_at -> Nullable<Timestamptz>,
-    }
-}
-
-table! {
-    origin_members (origin_id, account_id) {
-        origin_id -> BigInt,
-        origin_name -> Nullable<Text>,
-        account_id -> BigInt,
-        account_name -> Nullable<Text>,
-        created_at -> Nullable<Timestamptz>,
-        updated_at -> Nullable<Timestamptz>,
-    }
-}
-
-table! {
-    origin_packages (id) {
-        id -> BigInt,
-        origin_id -> BigInt,
-        owner_id -> BigInt,
-        name -> Text,
-        ident -> Text,
-        checksum -> Text,
-        manifest -> Text,
-        config -> Text,
-        target -> Text,
-        deps -> Text,
-        tdeps -> Text,
-        exposes -> Text,
-        visibility -> Text,
-        created_at -> Nullable<Timestamptz>,
-        updated_at -> Nullable<Timestamptz>,
-    }
-}
-
-table! {
-    origin_project_integrations (id) {
-        id -> BigInt,
-        origin -> Text,
-        name -> Text,
-        integration -> Text,
-        integration_name -> Text,
-        body -> Text,
-        project_id -> BigInt,
-        integration_id -> BigInt,
-        created_at -> Nullable<Timestamptz>,
-        updated_at -> Nullable<Timestamptz>,
-    }
-}
-
-table! {
-    origin_projects (id) {
-        id -> BigInt,
-        origin_id -> Nullable<BigInt>,
-        origin_name -> Nullable<Text>,
-        package_name -> Nullable<Text>,
-        name -> Nullable<Text>,
-        plan_path -> Nullable<Text>,
-        owner_id -> Nullable<BigInt>,
-        created_at -> Nullable<Timestamptz>,
-        updated_at -> Nullable<Timestamptz>,
-        visibility -> Text,
-    }
-}
-
-table! {
-    origin_public_keys (id) {
-        id -> BigInt,
-        origin_id -> BigInt,
-        owner_id -> BigInt,
-        name -> Text,
-        revision -> Text,
-        full_name -> Text,
-        body -> Bytea,
-        created_at -> Nullable<Timestamptz>,
-        updated_at -> Nullable<Timestamptz>,
-    }
-}
-
-table! {
-    origin_secret_keys (id) {
-        id -> BigInt,
-        origin_id -> BigInt,
-        owner_id -> BigInt,
-        name -> Text,
-        revision -> Text,
-        full_name -> Text,
-        body -> Bytea,
-        created_at -> Nullable<Timestamptz>,
-        updated_at -> Nullable<Timestamptz>,
-    }
-}
-
-table! {
-    origins (id) {
-        id -> BigInt,
-        name -> Text,
-        owner_id -> BigInt,
-        default_package_visibility -> Text,
-        created_at -> Nullable<Timestamptz>,
-        updated_at -> Nullable<Timestamptz>,
-    }
-}
