+++ components/builder-api/src/server/schema/key.rs
@@ -1,0 +1,27 @@
+table! {
+    origin_public_keys (id) {
+        id -> BigInt,
+        origin_id -> BigInt,
+        owner_id -> BigInt,
+        name -> Text,
+        revision -> Text,
+        full_name -> Text,
+        body -> Bytea,
+        created_at -> Nullable<Timestamptz>,
+        updated_at -> Nullable<Timestamptz>,
+    }
+}
+
+table! {
+    origin_secret_keys (id) {
+        id -> BigInt,
+        origin_id -> BigInt,
+        owner_id -> BigInt,
+        name -> Text,
+        revision -> Text,
+        full_name -> Text,
+        body -> Bytea,
+        created_at -> Nullable<Timestamptz>,
+        updated_at -> Nullable<Timestamptz>,
+    }
+}
