+++ components/builder-api/src/server/schema/origin.rs
@@ -1,0 +1,10 @@
+table! {
+    origins (id) {
+        id -> BigInt,
+        name -> Text,
+        owner_id -> BigInt,
+        default_package_visibility -> Text,
+        created_at -> Nullable<Timestamptz>,
+        updated_at -> Nullable<Timestamptz>,
+    }
+}
