+++ components/builder-originsrv/src/config.rs
@@ -30,11 +30,9 @@ pub struct Config {
 
 impl Default for Config {
     fn default() -> Self {
-        let mut datastore = DataStoreCfg::default();
-        datastore.database = String::from("builder");
         Config {
             app: AppCfg::default(),
-            datastore: datastore,
+            datastore: DataStoreCfg::default(),
             jobsrv_enabled: true,
         }
     }
